djadmin/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
psycopg2==2.9.3
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
